+++ govcd/lb_test.go
@@ -91,66 +91,6 @@ func (vcd *TestVCD) Test_LB(check *C) {
 	check.Assert(queryErr, IsNil)
 }
 
-func (vcd *TestVCD) createAngGetResourcesForVmCreation(check *C, vmName string) (*Vdc, *EdgeGateway, VAppTemplate, *VApp, types.NetworkConnectionSection, error) {
-	// Get org and vdc
-	org, err := vcd.client.GetAdminOrgByName(vcd.config.VCD.Org)
-	check.Assert(err, IsNil)
-	vdc, err := org.GetVDCByName(vcd.config.VCD.Vdc, false)
-	check.Assert(err, IsNil)
-	check.Assert(vdc, NotNil)
-	edge, err := vcd.vdc.GetEdgeGatewayByName(vcd.config.VCD.EdgeGateway, false)
-	check.Assert(err, IsNil)
-	// Find catalog and catalog item
-	catalog, err := org.GetCatalogByName(vcd.config.VCD.Catalog.Name, false)
-	check.Assert(err, IsNil)
-	check.Assert(catalog, NotNil)
-	catalogItem, err := catalog.GetCatalogItemByName(vcd.config.VCD.Catalog.CatalogItem, false)
-	check.Assert(err, IsNil)
-	// Skip the test if catalog item is not Photon OS
-	if !isItemPhotonOs(*catalogItem) {
-		check.Skip(fmt.Sprintf("Skipping test because catalog item %s is not Photon OS",
-			vcd.config.VCD.Catalog.CatalogItem))
-	}
-	fmt.Printf("# Creating RAW vApp '%s'", vmName)
-	vappTemplate, err := catalogItem.GetVAppTemplate()
-	check.Assert(err, IsNil)
-	// Compose Raw vApp
-	err = vdc.ComposeRawVApp(vmName)
-	check.Assert(err, IsNil)
-	vapp, err := vdc.GetVAppByName(vmName, true)
-	check.Assert(err, IsNil)
-	// vApp was created - let's add it to cleanup list
-	AddToCleanupList(vmName, "vapp", "", "createTestVapp")
-	// Wait until vApp becomes configurable
-	initialVappStatus, err := vapp.GetStatus()
-	check.Assert(err, IsNil)
-	if initialVappStatus != "RESOLVED" { // RESOLVED vApp is ready to accept operations
-		err = vapp.BlockWhileStatus(initialVappStatus, vapp.client.MaxRetryTimeout)
-		check.Assert(err, IsNil)
-	}
-	fmt.Printf(". Done\n")
-	fmt.Printf("# Attaching vDC network '%s' to vApp '%s'", vcd.config.VCD.Network.Net1, vmName)
-	// Attach vDC network to vApp so that VMs can use it
-	net, err := vdc.GetOrgVdcNetworkByName(vcd.config.VCD.Network.Net1, false)
-	check.Assert(err, IsNil)
-	task, err := vapp.AddRAWNetworkConfig([]*types.OrgVDCNetwork{net.OrgVDCNetwork})
-	check.Assert(err, IsNil)
-	err = task.WaitTaskCompletion()
-	check.Assert(err, IsNil)
-	fmt.Printf(". Done\n")
-	// Spawn 2 VMs with python servers in the newly created vApp
-	desiredNetConfig := types.NetworkConnectionSection{}
-	desiredNetConfig.PrimaryNetworkConnectionIndex = 0
-	desiredNetConfig.NetworkConnection = append(desiredNetConfig.NetworkConnection,
-		&types.NetworkConnection{
-			IsConnected:             true,
-			IPAddressAllocationMode: types.IPAllocationModePool,
-			Network:                 vcd.config.VCD.Network.Net1,
-			NetworkConnectionIndex:  0,
-		})
-	return vdc, edge, vappTemplate, vapp, desiredNetConfig, err
-}
-
 // validateTestLbPrerequisites verifies the following:
 // * Edge Gateway is set in config
 // * ExternalIp is set in config (will be edge gateway external IP)
